[FIX] html_builder: enable "On Hover" animation for images with shapes
Applying a shape to an image unintentionally disabled the "On Hover"
animation option. This commit restores the option with proper checks,
ensuring it remains disabled only when the applied shape includes an
animation.

The related code was also refactored slightly for improved readability.

Steps to reproduce:
1. Enter edit mode in Website.
2. Add an image and apply a shape.
3. Notice that the "On Hover" animation option is incorrectly hidden.
